Add networks from ironic and neutron to OpenstackInfra
Add networks to OpenstackInfra, moving cloud network management
into shared code. Joining data from ironic interfaces and
neutron as a preparation of showing the network of Hosts. But
that connection will need db change that will come as another PR
