+++ app/models/ems_openstack_infra.rb
@@ -1,6 +1,7 @@
 class EmsOpenstackInfra < EmsInfra
   include EmsOpenstackMixin
   include HasManyOrchestrationStackMixin
+  include HasManyCloudNetworksMixin
 
   before_save :ensure_parent_provider
   before_destroy :destroy_parent_provider
