+++ app/models/ems_refresh/save_inventory_infra.rb
@@ -52,7 +52,7 @@ module EmsRefresh::SaveInventoryInfra
     end
 
     child_keys = [:storages, :clusters, :hosts, :vms, :folders, :resource_pools, :customization_specs,
-                  :orchestration_templates, :orchestration_stacks]
+                  :orchestration_templates, :orchestration_stacks, :cloud_networks, :security_groups, :floating_ips]
 
     # Save and link other subsections
     save_child_inventory(ems, hashes, child_keys, target)
