+++ app/models/security_group_openstack_infra.rb
@@ -1,0 +1,2 @@
+class SecurityGroupOpenstackInfra < SecurityGroup
+end
